fix(frontend): split `API_URL` environment variable in two variables
Depending on the source of the request (server or client side), the URL of the API may not be the same. On server-side, since it may be used with docker compose, it may be `http://<docker_service_name>:<port>`. However, the client (the browser) doesn't know `http://<docker_service_name>`, so we need to use another URL.
This fix therefore allows to differentiate the two and permits to use the app with docker compose.